Restores Trait service
